Sync vertical new events and not rendered events
